requirements.txt
annotated-types==0.5.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==2.15.5
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.39.3
botocore==1.39.3
cachetools==5.3.1
certifi==2023.5.7
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
Clippy==0.6.4
coinbase-commerce==1.0.1
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==43.0.0
cssselect2==0.7.0
decorator==4.4.2
defusedxml==0.7.1
dek==1.1.0
dill==0.3.6
distlib==0.3.6
dj-database-url==2.0.0
Django==5.0
django-allauth==0.56.1
django-appconf==1.0.5
django-bootstrap==0.2.4
django-bootstrap-form==3.4
django-bootstrap5==24.2
django-braces==1.17.0
django-ckeditor==6.6.1
django-cors-headers==4.1.0
django-countries==7.5.1
django-crispy-forms==2.4
django-extensions==3.2.3
django-filter==25.1
django-heroku==0.3.1
django-hosts==6.0
django-js-asset==2.1.0
django-jsonresponse==0.10.0
django-model-utils==4.3.1
django-rest-framework==0.1.0
django-richtextfield==1.6.1
django-sites==0.11
django-ssl-auth==2.2.0
django-star-ratings==0.9.2
django-storages==1.13.2
django-uni-form==0.9.0
django-utils-six==2.0
djangorestframework==3.14.0
editor==1.3.0
env==0.1.0
filelock==3.13.1
flash==1.0.3
Flask==2.3.2
google-api-core==2.11.1
google-auth==2.23.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
isodate==0.6.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonfield==3.1.0
jsonpickle==3.0.1
lazy-object-proxy==1.9.0
lxml==5.3.0
manage.py==0.2.10
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector-python==8.1.0
numpy==2.3.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
pep8==1.7.1
pillow==11.2.1
platformdirs==3.8.0
proglog==0.1.10
protobuf==4.21.12
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pylint==2.17.4
pyparsing==3.1.0
pypdf==4.3.1
PyPDF2==3.0.1
pypi==2.1
pypng==0.20220715.0
pystache==0.6.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.9
runs==1.2.2
s3transfer==0.13.0
setuptools==80.9.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
static==1.1.1
svglib==1.5.1
swapper==1.3.0
texteditor==1.4.1
tinycss2==1.3.0
toml==0.10.2
tomlkit==0.11.8
tqdm==4.66.1
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==5.2
Unipath==1.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==1.26.16
virtualenv==20.23.1
virtualenvwrapper-win==1.2.7
webencodings==0.5.1
Werkzeug==2.3.6
whitenoise==6.5.0
wrapt==1.15.0
xhtml2pdf==0.2.16
xmod==1.4.0
zipp==3.15.0
